# config/ui/openapi.yaml
openapi: 3.1.0
info:
  title: mosdns Runtime API
  version: "2026-03-18"
  description: |
    当前规范只描述源码中真实注册的接口：
    - `/api/v1/*`
    - `/api/v3/audit/*`

    说明：
    - `/api/v1/control/overrides`、`/api/v1/control/upstreams`、`/api/v1/control/upstreams/tags` 是兼容别名，已在规范中标记为 `deprecated`
    - 核心 JSON 错误格式为 `{ "code": "...", "error": "..." }`
    - requery 插件接口使用独立错误格式 `{ "status": "error", "message": "..." }`
    - `GET /api/v1/update/status` 与 `POST /api/v1/update/check` 即使远程检查失败，当前实现仍返回 `200`
servers:
  - url: http://127.0.0.1:9099
    description: Default local control endpoint
...
paths:
...
  /api/v1/cache/{tag}/stats:
    parameters:
      - $ref: '#/components/parameters/TagPath'
    get:
      tags: [Cache]
      summary: Get one cache instance stats
      description: |
        返回单个缓存实例的运行态统计、最近 dump/load/WAL 状态和当前生效参数。

        说明：
        - 非持久化缓存会返回 `config.persist=false`
        - 非持久化缓存通常不会提供 `snapshot_file` / `wal_file`
        - `config.size` 表示最大缓存条目数，超过后按 LRU 淘汰旧条目
      responses:
        '200':
          description: Cache stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheStatsSnapshot'
        '404':
          $ref: '#/components/responses/ApiNotFound'
  /api/v1/cache/{tag}/entries:
    parameters:
      - $ref: '#/components/parameters/TagPath'
    get:
      tags: [Cache]
      summary: Query cache entries
      parameters:
        - $ref: '#/components/parameters/SearchQuery'
        - $ref: '#/components/parameters/OffsetQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Cache entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheEntriesResponse'
        '404':
          $ref: '#/components/responses/ApiNotFound'
        '500':
          $ref: '#/components/responses/ApiInternalError'
  /api/v1/cache/{tag}/save:
    parameters:
      - $ref: '#/components/parameters/TagPath'
    post:
      tags: [Cache]
      summary: Persist cache instance
      responses:
        '200':
          description: Saved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TagMessageResponse'
        '404':
          $ref: '#/components/responses/ApiNotFound'
        '500':
          $ref: '#/components/responses/ApiInternalError'
  /api/v1/cache/{tag}/flush:
    parameters:
      - $ref: '#/components/parameters/TagPath'
    post:
      tags: [Cache]
      summary: Flush cache instance
      responses:
        '200':
          description: Flushed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/ApiNotFound'
        '500':
          $ref: '#/components/responses/ApiInternalError'
  /api/v1/cache/{tag}/purge_domain:
    parameters:
      - $ref: '#/components/parameters/TagPath'
    post:
      tags: [Cache]
      summary: Purge cache by domain
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurgeDomainRequest'
      responses:
        '200':
          description: Purge result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurgeDomainResponse'
        '400':
          $ref: '#/components/responses/ApiBadRequest'
        '404':
          $ref: '#/components/responses/ApiNotFound'
  /api/v1/cache/purge_domains:
    post:
      tags: [Cache]
      summary: Purge domains from runtime caches
      description: |
        从运行中的响应缓存批量清理一组域名，可按缓存标签定向，也可选择同时包含 UDP fast cache。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PurgeDomainsRequest'
      responses:
        '200':
          description: Purge results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeCacheActionResponse'
        '400':
          $ref: '#/components/responses/ApiBadRequest'
        '404':
          $ref: '#/components/responses/ApiNotFound'
        '502':
          $ref: '#/components/responses/ApiBadGateway'
  /api/v1/cache/flush_all:
    post:
      tags: [Cache]
      summary: Flush runtime caches in batch
      description: |
        批量清空运行中的响应缓存，可按缓存标签定向，也可选择同时包含 UDP fast cache。
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlushAllRequest'
      responses:
        '200':
          description: Flush results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeCacheActionResponse'
        '400':
          $ref: '#/components/responses/ApiBadRequest'
        '404':
          $ref: '#/components/responses/ApiNotFound'
        '502':
          $ref: '#/components/responses/ApiBadGateway'
...
components:
  responses:
    ApiBadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIErrorResponse'
...
    ApiNotFound:
      description: Resource or plugin not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIErrorResponse'
...
    ApiInternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIErrorResponse'
...
    ApiBadGateway:
      description: Upstream or remote dependency failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIErrorResponse'
...
  parameters:
...
    LimitQuery:
      name: limit
      in: query
      description: 返回条数或上限。具体默认值由各接口决定。
      schema:
        type: integer
        minimum: 1
    OffsetQuery:
      name: offset
      in: query
      description: 偏移量
      schema:
        type: integer
        minimum: 0
...
    SearchQuery:
      name: q
      in: query
      description: 搜索关键字
      schema:
        type: string
...
    TagPath:
      name: tag
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    APIErrorResponse:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
        error:
          type: string
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    TagMessageResponse:
      type: object
      required: [message, tag]
      properties:
        message:
          type: string
        tag:
          type: string
...
    CacheOpStatus:
      type: object
      properties:
        status:
          type: string
          description: "`ok` / `error` / `not_run`"
        at:
          type: string
          format: date-time
        duration:
          type: string
        entries:
          type: integer
        error:
          type: string
    CacheStatsConfig:
      type: object
      description: 当前缓存实例实际生效的运行参数
      properties:
        size:
          type: integer
          description: 最大缓存条目数；达到上限后按 LRU 淘汰旧条目
        lazy_cache_ttl:
          type: integer
          description: TTL 过期后允许返回 stale 结果的窗口，单位秒
        l1_enabled:
          type: boolean
        l1_total_cap:
          type: integer
        l1_shard_cap:
          type: integer
        enable_ecs:
          type: boolean
        persist:
          type: boolean
          description: 是否启用 dump + WAL 持久化
        dump_interval:
          type: integer
          description: 快照周期，单位秒；仅 `persist=true` 时出现
        wal_sync_interval:
          type: integer
          description: WAL flush/fsync 周期，单位秒；仅 `persist=true` 时出现
      additionalProperties: true
    CacheStatsSnapshot:
      type: object
      description: 单个缓存实例的运行态快照
      properties:
        key:
          type: string
        name:
          type: string
        tag:
          type: string
        snapshot_file:
          type: string
        wal_file:
          type: string
        backend_size:
          type: integer
        l1_size:
          type: integer
        updated_keys:
          type: integer
          format: int64
        counters:
          type: object
          additionalProperties:
            type: integer
            format: int64
        last_dump:
          $ref: '#/components/schemas/CacheOpStatus'
        last_load:
          $ref: '#/components/schemas/CacheOpStatus'
        last_wal_replay:
          $ref: '#/components/schemas/CacheOpStatus'
        config:
          $ref: '#/components/schemas/CacheStatsConfig'
        error:
          type: string
...
    CacheEntry:
      type: object
      properties:
        key:
          type: string
        domain_set:
          type: string
        stored_time:
          type: string
        msg_expire:
          type: string
        cache_expire:
          type: string
        dns_message:
          type: string
    CacheEntriesResponse:
      type: object
      properties:
        tag:
          type: string
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/CacheEntry'
    PurgeDomainRequest:
      type: object
      required: [qname]
      properties:
        qname:
          type: string
        qtype:
          type: integer
    PurgeDomainResponse:
      type: object
      properties:
        qname:
          type: string
        qtype:
          type: integer
        purged:
          type: integer
    PurgeDomainsRequest:
      type: object
      required: [domains]
      properties:
        tags:
          type: array
          items:
            type: string
        domains:
          type: array
          items:
            type: string
        qtypes:
          type: array
          items:
            type: integer
        include_udp_fast:
          type: boolean
    FlushAllRequest:
      type: object
      properties:
        tags:
          type: array
          items:
            type: string
        include_udp_fast:
          type: boolean
    RuntimeCacheActionItem:
      type: object
      properties:
        tag:
          type: string
        kind:
          type: string
        ok:
          type: boolean
        purged:
          type: integer
        error:
          type: string
    RuntimeCacheActionResponse:
      type: object
      properties:
        success:
          type: integer
        failed:
          type: integer
        total:
          type: integer
        items:
          type: array
          items:
            $ref: '#/components/schemas/RuntimeCacheActionItem'
...